Introduce SavePostToDbUseCase
