+++ WordPress/src/main/java/org/wordpress/android/ui/posts/EditPostViewModel.kt
@@ -8,25 +8,17 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import org.wordpress.android.fluxc.Dispatcher
-import org.wordpress.android.fluxc.generated.PostActionBuilder
 import org.wordpress.android.fluxc.model.PostImmutableModel
 import org.wordpress.android.fluxc.model.PostModel
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.post.PostStatus
-import org.wordpress.android.fluxc.model.post.PostStatus.DRAFT
 import org.wordpress.android.fluxc.store.SiteStore
-import org.wordpress.android.modules.BG_THREAD
 import org.wordpress.android.modules.UI_THREAD
-import org.wordpress.android.ui.notifications.utils.PendingDraftsNotificationsUtilsWrapper
 import org.wordpress.android.ui.posts.EditPostViewModel.UpdateFromEditor.Failed
 import org.wordpress.android.ui.posts.EditPostViewModel.UpdateFromEditor.PostFields
 import org.wordpress.android.ui.posts.EditPostViewModel.UpdateResult.Error
 import org.wordpress.android.ui.posts.EditPostViewModel.UpdateResult.Success
 import org.wordpress.android.ui.uploads.UploadServiceFacade
-import org.wordpress.android.ui.uploads.UploadUtilsWrapper
 import org.wordpress.android.util.AppLog
-import org.wordpress.android.util.DateTimeUtilsWrapper
 import org.wordpress.android.util.NetworkUtilsWrapper
 import org.wordpress.android.viewmodel.Event
 import org.wordpress.android.viewmodel.ScopedViewModel
@@ -39,14 +31,10 @@ private const val MAX_UNSAVED_POSTS = 50
 class EditPostViewModel
 @Inject constructor(
     @Named(UI_THREAD) private val mainDispatcher: CoroutineDispatcher,
-    @Named(BG_THREAD) private val bgDispatcher: CoroutineDispatcher,
-    private val dispatcher: Dispatcher,
     private val siteStore: SiteStore,
-    private val uploadUtils: UploadUtilsWrapper,
     private val postUtils: PostUtilsWrapper,
-    private val pendingDraftsNotificationsUtils: PendingDraftsNotificationsUtilsWrapper,
     private val uploadService: UploadServiceFacade,
-    private val dateTimeUtils: DateTimeUtilsWrapper,
+    private val savePostToDbUseCase: SavePostToDbUseCase,
     private val networkUtils: NetworkUtilsWrapper
 ) : ScopedViewModel(mainDispatcher) {
     private var debounceCounter = 0
@@ -62,7 +50,7 @@ class EditPostViewModel
         editPostRepository: EditPostRepository,
         site: SiteModel
     ): Boolean {
-        savePostToDb(context, editPostRepository, site)
+        savePostToDbUseCase.savePostToDb(context, editPostRepository, site)
         return if (networkUtils.isNetworkAvailable()) {
             postUtils.trackSavePostAnalytics(
                     editPostRepository.getPost(),
@@ -100,55 +88,14 @@ class EditPostViewModel
             _onSavePostTriggered.value = Event(Unit)
         }
     }
+
     fun savePostToDb(
         context: Context,
         postRepository: EditPostRepository,
         site: SiteModel
     ) {
         Log.d("vojta", "Saving post to DB")
-        if (postRepository.postHasChangesFromDb()) {
-            val post = checkNotNull(postRepository.getEditablePost())
-            // mark as pending if the user doesn't have publishing rights
-            if (!uploadUtils.userCanPublish(site)) {
-                when (postRepository.status) {
-                    PostStatus.UNKNOWN,
-                    PostStatus.PUBLISHED,
-                    PostStatus.SCHEDULED,
-                    PostStatus.PRIVATE ->
-                        post.setStatus(PostStatus.PENDING.toString())
-                    PostStatus.DRAFT,
-                    PostStatus.PENDING,
-                    PostStatus.TRASHED -> {
-                    }
-                }
-            }
-            post.setIsLocallyChanged(true)
-            post.setDateLocallyChanged(dateTimeUtils.currentTimeInIso8601UTC())
-            handlePendingDraftNotifications(context, postRepository)
-            Log.d("vojta", "Post has changes so really saving")
-            postRepository.saveDbSnapshot()
-            dispatcher.dispatch(PostActionBuilder.newUpdatePostAction(post))
-        }
-    }
-
-    private fun handlePendingDraftNotifications(
-        context: Context,
-        editPostRepository: EditPostRepository
-    ) {
-        if (editPostRepository.status == DRAFT) {
-            // now set the pending notification alarm to be triggered in the next day, week, and month
-            pendingDraftsNotificationsUtils
-                    .scheduleNextNotifications(
-                            context,
-                            editPostRepository.id,
-                            editPostRepository.dateLocallyChanged
-                    )
-        } else {
-            pendingDraftsNotificationsUtils.cancelPendingDraftAlarms(
-                    context,
-                    editPostRepository.id
-            )
-        }
+        savePostToDbUseCase.savePostToDb(context, postRepository, site)
     }
 
     fun updatePostObjectWithUI(
